+++ bika/lims/upgrade/v01_03_000.py
@@ -5,8 +5,13 @@
 # Copyright 2018 by it's authors.
 # Some rights reserved. See LICENSE.rst, CONTRIBUTORS.rst.
 
+from Products.DCWorkflow.Guard import Guard
+from Products.ZCatalog.ProgressHandler import ZLogHandler
 from bika.lims import api
 from bika.lims import logger
+from bika.lims.catalog.analysis_catalog import CATALOG_ANALYSIS_LISTING
+from bika.lims.catalog.analysisrequest_catalog import \
+    CATALOG_ANALYSIS_REQUEST_LISTING
 from bika.lims.config import PROJECTNAME as product
 from bika.lims.upgrade import upgradestep
 from bika.lims.upgrade.utils import UpgradeUtils
@@ -59,6 +64,9 @@ def upgrade(tool):
     # https://github.com/senaite/senaite.core/pull/1060
     purge_portlets(portal)
 
+    # Setup the partitioning system
+    setup_partitioning(portal)
+
     # Fix Cannot get the allowed transitions (guard_sample_prep_transition)
     # https://github.com/senaite/senaite.core/pull/1069
     remove_sample_prep_workflow(portal)
@@ -102,6 +110,111 @@ def purge_portlets(portal):
     setup.runImportStepFromProfile(profile, 'portlets')
     logger.info("Purging portlets [DONE]")
 
+def setup_partitioning(portal):
+    """Setups the enhanced partitioning system
+    """
+    logger.info("Setting up the enhanced partitioning system")
+
+    # Add "Create partition" transition
+    add_create_partition_transition(portal)
+
+    # Add getAncestorsUIDs index in analyses catalog
+    add_partitioning_indexes(portal)
+
+    # Adds metadata columns for partitioning
+    add_partitioning_metadata(portal)
+
+
+def add_create_partition_transition(portal):
+    logger.info("Adding partitioning workflow")
+    wf_tool = api.get_tool("portal_workflow")
+    workflow = wf_tool.getWorkflowById("bika_ar_workflow")
+
+    # Transition: create_partitions
+    update_role_mappings = False
+    transition_id = "create_partitions"
+    if transition_id not in workflow.transitions:
+        update_role_mappings = True
+        workflow.transitions.addTransition(transition_id)
+    transition = workflow.transitions.create_partitions
+    transition.setProperties(
+        title="Create partitions",
+        new_state_id='sample_received',
+        after_script_name='',
+        actbox_name="Create partitions", )
+    guard = transition.guard or Guard()
+    guard_props = {'guard_permissions': 'BIKA: Edit Results',
+                   'guard_roles': '',
+                   'guard_expr': 'python:here.guard_handler("create_partitions")'}
+    guard.changeFromProperties(guard_props)
+    transition.guard = guard
+
+    # Add the transition to "sample_received" state
+    state = workflow.states.sample_received
+    if transition_id not in state.transitions:
+        update_role_mappings = True
+        state.transitions += (transition_id, )
+
+    if not update_role_mappings:
+        return
+
+    # Update role mappings if necessary. Analysis Requests transitioned beyond
+    # sample_received state are omitted.
+    states = ["sampled", "scheduled_sampling", "to_be_preserved",
+              "to_be_sampled", "sample_due", "sample_received"]
+    query = dict(portal_type="AnalysisRequest", states=states)
+    ars = api.search(query, CATALOG_ANALYSIS_REQUEST_LISTING)
+    total = len(ars)
+    for num, ar in enumerate(ars, start=1):
+        workflow.updateRoleMappingsFor(api.get_object(ar))
+        if num % 100 == 0:
+            logger.info("Updating role mappings: {0}/{1}".format(num, total))
+
+
+def add_partitioning_indexes(portal):
+    """Adds the indexes for partitioning
+    """
+    logger.info("Adding partitioning indexes")
+    add_index(portal, catalog_id=CATALOG_ANALYSIS_LISTING,
+              index_name="getAncestorsUIDs",
+              index_attribute="getAncestorsUIDs",
+              index_metatype="KeywordIndex")
+
+
+def add_partitioning_metadata(portal):
+    """Add metadata columns required for partitioning machinery
+    """
+    logger.info("Adding partitioning metadata")
+    add_metadata(portal, CATALOG_ANALYSIS_REQUEST_LISTING,
+                 'getRawParentAnalysisRequest')
+
+
+def add_index(portal, catalog_id, index_name, index_attribute, index_metatype):
+    logger.info("Adding '{}' index to '{}' ...".format(index_name, catalog_id))
+    catalog = api.get_tool(catalog_id)
+    if index_name in catalog.indexes():
+        logger.info("Index '{}' already in catalog '{}' [SKIP]"
+                    .format(index_name, catalog_id))
+        return
+    catalog.addIndex(index_name, index_metatype)
+    logger.info("Indexing new index '{}' ...".format(index_name))
+    catalog.manage_reindexIndex(index_name)
+
+
+def add_metadata(portal, catalog_id, column, refresh_catalog=False):
+    logger.info("Adding '{}' metadata to '{}' ...".format(column, catalog_id))
+    catalog = api.get_tool(catalog_id)
+    if column in catalog.schema():
+        logger.info("Metadata '{}' already in catalog '{}' [SKIP]"
+                    .format(column, catalog_id))
+        return
+
+    catalog.addColumn(column)
+
+    if refresh_catalog:
+        logger.info("Refreshing catalog '{}' ...".format(catalog_id))
+        handler = ZLogHandler(steps=100)
+        catalog.refreshCatalog(pghandler=handler)
 
 def remove_sample_prep_workflow(portal):
     """Removes sample_prep and sample_prep_complete transitions